In tests: Use $::env(DESIGN_TOP) where it isn't already.
This makes all tests use the same concept for accessing the
test verilog file in question. It also makes it possible to
run tests from different directories if only DESIGN_TOP is set
correctly.

In one place, a filename is resolved intended relative to the script,
but currently requires the script to be executed in the same
directory; make that location independent by reading it
relative to [file dirname [info script]]

Signed-off-by: Henner Zeller <h.zeller@acm.org>