Make output path for test outputs configurable.
In preparation for CI's and other test environments in which
the test outputs should be stored out-of-tree.

The new file test-utils.tcl provides a function that determines
the test output path (currently just by concatenating the filename
with a path from the environment). The Makefile_test.common sources
this utils file and the corresponding test file.

Since this adds a convenient test-utils.tcl file, this might also
be useful to put some testing boilerplate functionality there in
the future.

Signed-off-by: Henner Zeller <h.zeller@acm.org>