Add .editorconfig, to help contributors honor coding style.
Setting the indentation for tcl and cc to 4 characters which
is mostly what is happening in this repository. There are
a few tabs here and there, but they will vanish with changes
happening there

The .editorconfig file is supported by a vast number of IDEs and
editors, so hopefully this will make it simpler for would-be
contributors.

Signed-off-by: Henner Zeller <h.zeller@acm.org>