Fix loading of ql-qlf-k4n8 plugin in tests; use canonical test.v name
The line that is loading the plug-in is meant to skip attempting
loading a plugin with a particular name if the symbol we're interested
in is already there.

So in the ql-qlf-k4n8 case, we have to test for synth_quicklogic:

-if { [info procs ql-qlf-k4n8] == {} } { plugin -i ql-qlf-k4n8 }
+if { [info procs synth_quicklogic] == {} } { plugin -i ql-qlf-k4n8 }

(reasaon for all that: this allows to statically link the plugins
and have these scripts work idempotently).

For loading the test-related *.v files: use `$::env(DESIGN_TOP).v`
instead of the direct name. That allows it to run independently
from where the tests are invoked, as the test-driver can set the
environment variable to the fully qualified path.

Signed-off-by: Henner Zeller <h.zeller@acm.org>